requirements.txt
langchain>=1.0.0
langchain-openai>=0.2.0
langgraph>=1.0.0
langfuse>=2.0.0
boto3>=1.34.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pydantic>=2.8.0
python-dotenv>=1.0.1
requests>=2.32.0
pandas>=2.2.0
openpyxl>=3.1.0
pytest>=8.0.0
pytest-mock>=3.14.0
